pyproject.toml
[project]
name = "app"
version = "0.7.1"
description = "Сайт книг для IT специалистов"
authors = [{ name = "ig-rudenko", email = "ig.rudenko1@yandex.ru" }]
requires-python = ">=3.11,<4.0"
readme = "README.md"
dependencies = [
    "fastapi>=0.115.12,<0.116",
    "uvicorn>=0.34.3,<0.35",
    "sqlalchemy[asyncio]>=2.0.41,<3",
    "alembic>=1.16.1,<2",
    "passlib>=1.7.4,<2",
    "pydantic[email]>=2.11.5,<3",
    "python-jose[cryptography]>=3.5.0,<4",
    "bcrypt>=4.2.1,<5",
    "pillow>=11.2.1,<12",
    "python-multipart>=0.0.20,<0.0.21",
    "PyMuPDF>=1.24.10,<2",
    "pydantic-settings>=2.9.1,<3",
    "asyncpg>=0.30.0,<0.31",
    "aiofiles>=24.1.0,<25",
    "python-slugify>=8.0.4,<9",
    "celery[redis]>=5.5.3,<6",
    "aiohttp>=3.12.12,<4",
    "loguru>=0.7.3,<0.8",
    "aioboto3>=14.3.0,<15",
    "pyjwt>=2.10.1",
    "advanced-alchemy>=1.6.3",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0,<0.22",
    "httpx>=0.28.1,<0.29",
    "coverage>=7.8.0,<8",
    "black[d]>=25.1.0,<26",
    "mypy>=1.15.0,<2",
    "types-passlib>=1.7.7.20240327,<2",
    "types-python-jose>=3.3.4.20240106,<4",
    "types-aiofiles>=23.2.0.20240403,<24",
    "celery-types>=0.23.0,<0.24",
    "pydeps>=1.12.20,<2",
    "graphviz>=0.20.3,<0.21",
]

[tool.uv]
default-groups = [
    "dev",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
lint.flake8-bugbear.extend-immutable-calls = ["Depends", "fastapi.Depends", "fastapi.params.Depends"]
lint.extend-select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
lint.ignore = [
    # Line too long
    "E501",
    # Functions with a high complexity
    "C901",
    # Use ternary operator instead of `if`-`else`-block
    "SIM108",
    # Module level import not at top of file
    "E402"
]